ANDROID: fix binder change in merge of 4.4.183
The 4.4.183 merge was missing the change to the
binder driver associated with the linux-4.4.y
commit 8f6345a11caa ("coredump: fix race condition
between mmget_not_zero()/get_task_mm() and core dumping").
It was left out because the android-4.4 binder driver
has been significantly refactored compared to linux-4.4.y.

This patch applies the missing change from that
patch to the binder driver.

Change-Id: I803e558cb629fedff04b14e23ac9f83e98628ede
Fixes: 94fd42809f7d ("Merge 4.4.183 into android-4.4")
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>